+++ move.php
@@ -27,7 +27,7 @@ if (check_login ($db, $lang, $langvars)) // Checks player login, sets playerinfo
 $title = $langvars['l_move_title'];
 include './header.php';
 
-$sector  = filter_input (INPUT_GET, 'sector', FILTER_SANITIZE_NUMBER_INT);
+$sector  = (int) filter_input (INPUT_GET, 'sector', FILTER_SANITIZE_NUMBER_INT);
 
 // Database driven language entries
 $langvars = BntTranslate::load ($db, $lang, array ('move', 'common', 'global_includes', 'global_funcs', 'combat', 'footer', 'news'));
